FoetoPath_1.2/requirements.txt
flask>=3.0
openslide-python>=1.3
openslide-bin>=4.0
Pillow>=10.0
